requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.7
Django==4.2.7
djangorestframework==3.14.0
jwt==1.3.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
